Create a crate that allows the use of init_tracing_for_test par any crate
