+++ crates/matrix-sdk-crypto/Cargo.toml
@@ -88,6 +88,7 @@ http.workspace = true
 indoc = "2.0.5"
 insta.workspace = true
 matrix-sdk-test.workspace = true
+matrix-sdk-test-utils.workspace = true
 proptest.workspace = true
 rmp-serde.workspace = true
 similar-asserts.workspace = true
