+++ crates/matrix-sdk-ui/Cargo.toml
@@ -69,6 +69,7 @@ assert_matches2.workspace = true
 eyeball-im-util.workspace = true
 matrix-sdk = { workspace = true, features = ["testing", "sqlite"] }
 matrix-sdk-test.workspace = true
+matrix-sdk-test-utils.workspace = true
 stream_assert.workspace = true
 tempfile.workspace = true
 url.workspace = true
