+++ crates/matrix-sdk-ui/src/lib.rs
@@ -31,7 +31,7 @@ pub use self::{room_list_service::RoomListService, timeline::Timeline};
 const DEFAULT_SANITIZER_MODE: HtmlSanitizerMode = HtmlSanitizerMode::Compat;
 
 #[cfg(test)]
-matrix_sdk_test::init_tracing_for_tests!();
+matrix_sdk_test_utils::init_tracing_for_tests!();
 
 #[cfg(feature = "uniffi")]
 uniffi::setup_scaffolding!();
