+++ crates/matrix-sdk-ui/tests/integration/main.rs
@@ -29,7 +29,7 @@ mod sliding_sync;
 mod sync_service;
 mod timeline;
 
-matrix_sdk_test::init_tracing_for_tests!();
+matrix_sdk_test_utils::init_tracing_for_tests!();
 
 /// Mount a Mock on the given server to handle the `GET /sync` endpoint with
 /// an optional `since` param that returns a 200 status code with the given
