+++ crates/matrix-sdk/src/lib.rs
@@ -98,4 +98,4 @@ pub mod live_location_share;
 pub mod test_utils;
 
 #[cfg(test)]
-matrix_sdk_test::init_tracing_for_tests!();
+matrix_sdk_test_utils::init_tracing_for_tests!();
