+++ crates/matrix-sdk/tests/integration/main.rs
@@ -22,7 +22,7 @@ mod send_queue;
 #[cfg(feature = "experimental-widgets")]
 mod widget;
 
-matrix_sdk_test::init_tracing_for_tests!();
+matrix_sdk_test_utils::init_tracing_for_tests!();
 
 /// Mount a Mock on the given server to handle the `GET /sync` endpoint with
 /// an optional `since` param that returns a 200 status code with the given
