+++ testing/matrix-sdk-integration-testing/src/lib.rs
@@ -1,6 +1,6 @@
 #![cfg(test)]
 
-matrix_sdk_test::init_tracing_for_tests!();
+matrix_sdk_test_utils::init_tracing_for_tests!();
 
 mod helpers;
 mod tests;
