+++ testing/matrix-sdk-test-utils/README.md
@@ -1,0 +1,1 @@
+ Utils function for the test helpers
